1st Data Analytics/Viz Project
I did step up in order for me to put all what I've got after a month long training in DataCamp (props to Developers Circles of Facebook for giving me an opportunity to learn in the platform). That's I did so far.. looking forward to improve the visuals and also the analysis. I am looking forward to improve and work with other projects. 

Thanks!